// client/src/app/_models/offer.ts
export class Offer {
    id: number;
    emp_id: number;
    name: string;
    company: number;
    consumption: number;
    pfc: number;
    pfc_market: number;
    profile: number;
    cc: number;
    years: Array<number>;
    lissage_years: Object;
    lissage: boolean;
    unit: string;
    offer_type: string;
    riscs: Array<number>;
    shedules: Array<number>;
    validation_time: number;
    employees: string;
    meters: Array<string>;
    decotes: Object;
    efforts: Object;
    energies: Object;
    majors: Object;
    ps1: Object;
    ps2: Object;
    user: number;
    second_type: String;
    showStatusPopup = false;
    signed_file: String;
    unsigned_file: String;
    nr_opportunite: number;
    comment: String;
    energy_type: string;
    date_debut: string;
    expiration_date: string;
    heures_in: string;
    date_fin: string;
    heures_fin: string;
    durre: string;
    sur_go: Object;
    grd: number;
    marche: boolean;
    releve: boolean;
    offer_status: string;
    conseiller: number;
    entreprise: string;
    contact: string;
    nomEntreprice: string;
    lis_years: string;
    percent: number;
    volumetrie: number;
    fonction: number;
    marge: number;
    custom_eco: string;
    duree: any;
    pfc_date_first: any;
    pfc_date_last: any;
    years_liss_list: any;
    lissage_years_dic: any;
    expirated: boolean;
}